Fix some test failures for Dart for the Web
`currentDartPackagePath` won't work for the web, where paths are
transformed when compiled to JavaScript.  Disable the test for
browsers.

The `sortWithKey` test did not show much difference when run in a
browser (possibly due to the overhead being more expensive), and
sometimes was very slightly slower than the normal `sort`.  Make
the test use `BigInt`s instead of `int`s to make comparisons more
expensive so that performance benefits from `sortWithKey` are more
pronounced.