Add `timeAsyncOperation`
Add a `timeAsyncOperation` function as an asynchronous analogue of
`timeOperation`.  I considered instead making `timeOperation`
generic and making it return a `FutureOr` so that there could be only
a single implementation, but that would be more work for callers.
The implementations of `timeOperation`/`timeAsyncOperation` are
fairly trivial, so the duplication is not a big deal.