Update to Dart 3.8, update dependencies, update lints, reformat
Use the new Dart 3.8 formatter option to preserve trailing
commas.  I want that to apply only to this package and don't
want to force that on downstream consumers, so change the
package-level `analysis_options.yaml` from a symlink to its
own `.yaml` file.

Unfortunately this seems to require making a breaking change:
`dart format` appears to honor the `analysis_options.yaml`
file in `lib/` instead of the one in the package directory,
so rename the published one to `recommended.yaml` instead.